+++ package.json
@@ -11,9 +11,11 @@
   "dependencies": {
     "axios": "^1.3.5",
     "cheerio": "^1.0.0-rc.12",
-    "epub-gen": "^0.1.0",
+    "epub-check": "^0.3.0",
+    "epub-gen": "vuongpsht/epub-gen",
     "fs": "0.0.1-security",
     "lodash": "^4.17.21",
-    "omnia-ebook-generator": "^0.2.1"
+    "omnia-ebook-generator": "^0.2.1",
+    "sharp": "^0.32.0"
   }
 }
